clarified why transport is pipes and not ws
