+++ src/Artemis.Plugins.Modules.Discord/DiscordModule.cs
@@ -151,7 +151,7 @@ public class DiscordModule : Module<DiscordDataModel>
 
             //Subscribe to these events as well
             await _discordClient.SubscribeAsync(DiscordRpcEvent.VOICE_SETTINGS_UPDATE);
-            //await _discordClient.SubscribeAsync(DiscordRpcEvent.NOTIFICATION_CREATE);
+            await _discordClient.SubscribeAsync(DiscordRpcEvent.NOTIFICATION_CREATE);
             await _discordClient.SubscribeAsync(DiscordRpcEvent.VOICE_CONNECTION_STATUS);
             await _discordClient.SubscribeAsync(DiscordRpcEvent.VOICE_CHANNEL_SELECT);
         }
